requirements.txt
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.2
tenacity==9.0.0
python-telegram-bot==21.4
rich==13.7.1
fastapi==0.111.1
uvicorn==0.30.5
jinja2==3.1.4



